feat: hoist mariadb fixes in mysql
